+++ finance/companies/sovereigns.py
@@ -78,6 +78,8 @@ class SovereignStore(headstore.HeadStore):
                 sec_store.is_esg_labeled("Sustainable")
             elif labeled_bond_tag == "Labeled Sustainable Linked":
                 sec_store.is_esg_labeled("Sustainability-Linked Bonds")
+            elif labeled_bond_tag == "Labeled Sustainable/Sustainable Linked":
+                sec_store.is_esg_labeled("Sustainable/Sustainability-Linked Bonds")
 
             elif sustainability_tag == "Y*":
                 sec_store.is_sustainable()