+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - make pathes working in ML Azure
 - BClass replace N/A's
+- add Labeled Sustainable/Sustainable Linked to labeled bonds
 ### Changed
 - Split Excluded SClass Labels into Poor Data, Poor Score, and Excluded Sector
 ### Removed